+++ CookieClijkstra/Path.cs
@@ -3,11 +3,11 @@
     public struct Path
     {
         public Vertex To;
-        public double Distance;
+        public float Distance;
 
         public string Name;
 
-        public Path(Vertex to, double distance, string name)
+        public Path(Vertex to, float distance, string name)
         {
             this.To = to;
             this.Distance = distance;
